Move the generic handlers to another package so they are easier to find.
