+++ src/main/java/com/iopipe/generic/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This package contains generic entry points.
+ *
+ * @since 2018/08/13
+ */
+
+package com.iopipe.generic;
